authentication

+++ src/main/java/com/example/blog/controller/ArticleController.java
@@ -7,6 +7,7 @@ import com.example.blog.repository.ArticleRepository;
 import com.example.blog.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
@@ -62,6 +63,16 @@ public class ArticleController {
             return "redirect:/";
         }
 
+        if (!(SecurityContextHolder.getContext().getAuthentication() instanceof AnonymousAuthenticationToken))
+        {
+           UserDetails principal = (UserDetails) SecurityContextHolder.getContext()
+                   .getAuthentication().getPrincipal();
+
+           User entityUser = this.userRepository.findByEmail(principal.getUsername());
+
+           model.addAttribute("user", entityUser);
+        }
+
         Article article = this.articleRepository.findById(id).orElse(null);
 
         model.addAttribute("article", article);
@@ -82,6 +93,11 @@ public class ArticleController {
 
         Article article = this.articleRepository.findById(id).orElse(null);
 
+        if(!isUserAuthorOrAdmin(article))
+        {
+            return "redirect:/article/" + id;
+        }
+
         model.addAttribute("view", "article/edit");
         model.addAttribute("article", article);
 
@@ -99,6 +115,11 @@ public class ArticleController {
         }
 
         Article article = this.articleRepository.findById(id).orElse(null);
+
+        if(!isUserAuthorOrAdmin(article))
+        {
+            return "redirect:/article/" + id;
+        }
 
         article.setContent(articleBindingModel.getContent());
         article.setTitle(articleBindingModel.getTitle());
@@ -120,6 +141,11 @@ public class ArticleController {
 
         Article article = this.articleRepository.findById(id).orElse(null);
 
+        if(!isUserAuthorOrAdmin(article))
+        {
+            return "redirect:/article/" + id;
+        }
+
         model.addAttribute("article", article);
         model.addAttribute("view", "article/delete");
 
@@ -138,9 +164,26 @@ public class ArticleController {
 
         Article article = this.articleRepository.findById(id).orElse(null);
 
+        if(!isUserAuthorOrAdmin(article))
+        {
+            return "redirect:/article/" + id;
+        }
+
         //this removes the article from the database
         this.articleRepository.delete(article);
 
         return "redirect:/";
     }
+
+    // Checks if the current user is admin or author
+    private boolean isUserAuthorOrAdmin(Article article)
+    {
+        UserDetails user = (UserDetails) SecurityContextHolder.getContext()
+                .getAuthentication().getPrincipal();
+
+        User userEntity = this.userRepository.findByEmail(user.getUsername());
+
+        return userEntity.isAdmin() || userEntity.isAuthor(article);
+    }
+
 }
